+++ lib/src/util/preferences/preferences_storage.dart
@@ -4,8 +4,8 @@ import 'package:shared_preferences/shared_preferences.dart';
 /// Utility class to facilitate storing shared preferences
 abstract class SharedPreferenceStorage implements SimpleKeyValueStorage {
   /// Create a new instance of SharedPreferenceStorage
-  factory SharedPreferenceStorage(SharedPreferences preferences) =>
-      _SharedPreferenceStorage(preferences);
+  factory SharedPreferenceStorage(Future<SharedPreferences> preferences) =
+      _SharedPreferenceStorage;
 
   /// Saves the given string [value] to the storage with [key]
   Future<void> saveString({required String key, required String value});
@@ -21,73 +21,77 @@ abstract class SharedPreferenceStorage implements SimpleKeyValueStorage {
 
   /// Retrieves the stored string value for [key].
   /// Returns null if the value is not found
-  String? getString(String key);
+  Future<String?> getString(String key);
 
   /// Retrieves the stored boolean value for [key].
   /// Returns null if the value is not found
-  bool? getBoolean(String key);
+  Future<bool?> getBoolean(String key);
 
   /// Retrieves the stored integer value for [key].
   /// Returns null if the value is not found
-  int? getInt(String key);
+  Future<int?> getInt(String key);
 
   /// Retrieves the stored double value for [key].
   /// Returns null if the value is not found
-  double? getDouble(String key);
+  Future<double?> getDouble(String key);
 
   /// Deletes the value stored for [key]
   Future<void> deleteKey(String key);
 
   /// Returns true if there is a value for [key]
-  bool containsKey(String key);
+  Future<bool> containsKey(String key);
 }
 
 class _SharedPreferenceStorage implements SharedPreferenceStorage {
-  final SharedPreferences _sharedPreferences;
+  final Future<SharedPreferences> _sharedPreferences;
 
   _SharedPreferenceStorage(this._sharedPreferences);
 
   @override
   Future<void> saveString({required String key, required String value}) =>
-      _sharedPreferences.setString(key, value);
+      _sharedPreferences.then((prefs) => prefs.setString(key, value));
 
   @override
   Future<void> saveBoolean({required String key, required bool value}) =>
-      _sharedPreferences.setBool(key, value);
+      _sharedPreferences.then((prefs) => prefs..setBool(key, value));
 
   @override
   Future<void> saveInt({required String key, required int value}) =>
-      _sharedPreferences.setInt(key, value);
+      _sharedPreferences.then((prefs) => prefs..setInt(key, value));
 
   @override
   Future<void> saveDouble({required String key, required double value}) =>
-      _sharedPreferences.setDouble(key, value);
+      _sharedPreferences.then((prefs) => prefs.setDouble(key, value));
 
   @override
-  String? getString(String key) => _sharedPreferences.getString(key);
+  Future<String?> getString(String key) =>
+      _sharedPreferences.then((prefs) => prefs.getString(key));
 
   @override
-  bool? getBoolean(String key) => _sharedPreferences.getBool(key);
+  Future<bool?> getBoolean(String key) =>
+      _sharedPreferences.then((prefs) => prefs.getBool(key));
 
   @override
-  int? getInt(String key) => _sharedPreferences.getInt(key);
+  Future<int?> getInt(String key) =>
+      _sharedPreferences.then((prefs) => prefs.getInt(key));
 
   @override
-  double? getDouble(String key) => _sharedPreferences.getDouble(key);
+  Future<double?> getDouble(String key) =>
+      _sharedPreferences.then((prefs) => prefs.getDouble(key));
 
   @override
-  Future<void> deleteKey(String key) => _sharedPreferences.remove(key);
+  Future<void> deleteKey(String key) =>
+      _sharedPreferences.then((prefs) => prefs.remove(key));
 
   @override
-  bool containsKey(String key) => _sharedPreferences.containsKey(key);
+  Future<bool> containsKey(String key) =>
+      _sharedPreferences.then((prefs) => prefs.containsKey(key));
 
   @override
-  Future<String?> getValue({required String key}) =>
-      Future.value(getString(key));
+  Future<String?> getValue({required String key}) => getString(key);
 
   @override
-  Future<bool> hasValue({required String key}) =>
-      Future.value(containsKey(key));
+  Future<bool> hasValue({required String key}) => containsKey(key);
 
   @override
   Future<void> removeValue({required String key}) => deleteKey(key);